update page repository,optimize app.js
