+++ app/Http/Repositories/PostRepository.php
@@ -8,16 +8,11 @@
 
 namespace App\Http\Repositories;
 
-use App\Configuration;
 use App\Post;
 use App\Tag;
 use Carbon\Carbon;
-use DOMDocument;
-use DOMXPath;
-use function GuzzleHttp\Promise\queue;
 use Illuminate\Http\Request;
 use Lufficc\MarkDownParser;
-use PhpParser\NodeTraverser;
 
 /**
  * design for cache
@@ -177,9 +172,7 @@ class PostRepository extends Repository
             $request['published_at'] = Carbon::now();
         }
 
-        $html_content = $this->markDownParser->parse($request->get('content'), false);
-        $html_content = $this->convertHtml($html_content);
-
+        $html_content = $this->markDownParser->parse($request->get('content'), false, true);
         $post = auth()->user()->posts()->create(
             array_merge(
                 $request->except(['_token', 'description']),
@@ -223,8 +216,7 @@ class PostRepository extends Repository
 
         $post->saveConfig($request->all());
 
-        $html_content = $this->markDownParser->parse($request->get('content'), false);
-        $html_content = $this->convertHtml($html_content);
+        $html_content = $this->markDownParser->parse($request->get('content'), false, true);
         return $post->update(
             array_merge(
                 $request->except(['_token', 'description']),
@@ -233,99 +225,6 @@ class PostRepository extends Repository
                     'description' => $this->markDownParser->parse($request->get('description'), false),
                 ]
             ));
-    }
-
-    /**
-     * convert
-     * <div class="figure **" caption="caption">
-     *     <img ..>
-     *     <img ..>
-     *     ...
-     * </div>
-     * to
-     * <figure class="**">
-     *     <p><img ..></p>
-     *     <p><img ..></p>
-     *     ...
-     *     <figcaption>$caption</figcaption>
-     * </figure>
-     * @param DOMDocument $dom
-     * @return bool
-     */
-    private function parseDiv(DOMDocument $dom)
-    {
-        $xpath = new DOMXpath($dom);
-        $galleries = $xpath->query('//div[contains(@class, "figure")]');
-        $changed = false;
-        foreach ($galleries as $gallery) {
-            $figure = $dom->createElement('figure');
-            $figure->setAttribute('class', trim(str_replace('figure', '', $gallery->getAttribute('class'))));
-            $frag = $dom->createDocumentFragment();
-            $alt = '';
-            foreach ($xpath->query('p/img', $gallery) as $image) {
-                if (!$alt)
-                    $alt = $image->getAttribute('alt');
-                //wrapped with div
-                $div = $dom->createElement('div');
-                $div->appendChild($image);
-                $frag->appendChild($div);
-            }
-            //empty string if no attribute with the given name is found.
-            $caption = $gallery->getAttribute('caption');
-            if (!$caption)
-                $caption = $alt;
-            $frag->appendXML("<figcaption>$caption</figcaption>");
-            $figure->appendChild($frag);
-            $gallery->parentNode->replaceChild($figure, $gallery);
-            $changed = true;
-        }
-        return $changed;
-    }
-
-    /**
-     * convert
-     * <img .. alt='alt' class='figure'>
-     * to
-     * <figure>
-     *     <img .. alt='alt'>
-     *     <figcaption>$alt</figcaption>
-     * </figure>
-     * @param DOMDocument $dom
-     * @return bool
-     */
-    private function parseImage(DOMDocument $dom)
-    {
-        $xpath = new DOMXpath($dom);
-        $images = $xpath->query('//img[contains(@class, "figure")]');
-        $changed = false;
-        foreach ($images as $image) {
-            $src = $image->getAttribute('src');
-            $alt = $image->getAttribute('alt');
-            $figure = $dom->createElement('figure');
-            $frag = $dom->createDocumentFragment(); // create fragment
-            $imgNode = $dom->createElement('img');
-            $imgNode->setAttribute('src', $src);
-            $imgNode->setAttribute('alt', $alt);
-            $frag->appendChild($imgNode);
-            $frag->appendXML("<figcaption>$alt</figcaption>");
-            $figure->appendChild($frag);
-            $image->parentNode->replaceChild($figure, $image);
-            $changed = true;
-        }
-        return $changed;
-
-    }
-
-    private function convertHtml($html)
-    {
-        $dom = new DOMDocument();
-        $dom->loadHTML(mb_convert_encoding($html, 'HTML-ENTITIES', 'UTF-8'));
-        $changed = $this->parseImage($dom);
-        $changed = $this->parseDiv($dom) || $changed;
-        if ($changed) {
-            $html = $dom->saveHTML();
-        }
-        return $html;
     }
 
     public function tag()
